move the getDeployParams helper function to a new file
